refactored and split screens and error page
