+++ haterfrontend/src/App.js
@@ -1,30 +1,18 @@
-import "./App.css";
-import { useEffect, useState } from "react";
-import Profilebutton from "./components/profilebutton/Profilebutton.js";
-import Profiletab from "./components/profiletab/Profiletab.js";
+import Landing from "./screens/Landing.jsx";
 import Home from "./screens/Home.jsx";
+import ErrorPage from "./screens/Error";
+import React from "react";
+import { Routes, Route } from "react-router-dom";
 
 function App() {
-  const [show, setShow] = useState({ display: "none" });
-  const closeBox = () => {
-    setShow({ display: "none" });
-  };
-  const handleClick = () => {
-    setShow({ display: "flex" });
-  };
   return (
-    <div className="fullDiv">
-      <Profilebutton click={handleClick} />
-      <Profiletab
-        name={"verybamboo"}
-        tweets={114}
-        likes={426}
-        click={closeBox}
-        show={show}
-      />
-      <div className="mid"></div>
-      <div className="right">3</div>
-    </div>
+    <>
+      <Routes>
+        <Route path="/" element={<Landing></Landing>}></Route>
+        <Route path="/home" element={<Home></Home>}></Route>
+        <Route path="*" element={<ErrorPage></ErrorPage>}></Route>
+      </Routes>
+    </>
   );
 }
