+++ haterfrontend/src/screens/Error.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+function ErrorScreen() {
+  return (
+    <div>
+      <h1>ERROR SCREEN</h1>
+    </div>
+  );
+}
+
+export default ErrorScreen;
