+++ haterfrontend/src/screens/Home.jsx
@@ -1,56 +1,30 @@
-import React from "react";
-import Button from "@mui/material/Button";
-export default function Home() {
+import "../styles/Home.css";
+import { useEffect, useState } from "react";
+import Profilebutton from "../components/profilebutton/Profilebutton.js";
+import Profiletab from "../components/profiletab/Profiletab.js";
+
+function Home() {
+  const [show, setShow] = useState({ display: "none" });
+  const closeBox = () => {
+    setShow({ display: "none" });
+  };
+  const handleClick = () => {
+    setShow({ display: "flex" });
+  };
   return (
-    <div className="home">
-      <div className="left">
-        <img
-          className="image"
-          src={
-            "https://icons.iconarchive.com/icons/femfoyou/angry-birds/1024/angry-bird-icon.png"
-          }
-          alt="image"
-        />
-        <h3>Hate your interest</h3>
-        <h3>Hear what people are hating</h3>
-        <h3>Join the conversation</h3>
-      </div>
-      <div className="right">
-        <form className="form">
-          <input
-            className="input"
-            placeholder="Phone, email, or username:"
-            type="text"
-            name="name"
-          />
-
-          <input
-            className="input"
-            placeholder="Password "
-            type="text"
-            name="password"
-          />
-          <Button id="login" type="submit" variant="outlined">
-            Login
-          </Button>
-        </form>
-
-        <div className="middle-right">
-          <h3>See what's happening in the world right now</h3>
-          <h4>Join Hater Today</h4>
-          <Button id="sign-up" variant="outlined">
-            Sign up with Google
-          </Button>
-
-          <Button id="sign-up" variant="outlined">
-            Sign up with phone or email
-          </Button>
-
-          {/* <Button id="main-login" variant="outlined">
-            Login
-          </Button> */}
-        </div>
-      </div>
+    <div className="fullDiv">
+      <Profilebutton click={handleClick} />
+      <Profiletab
+        name={"verybamboo"}
+        tweets={114}
+        likes={426}
+        click={closeBox}
+        show={show}
+      />
+      <div className="mid"></div>
+      <div className="right">3</div>
     </div>
   );
 }
+
+export default Home;
